backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "saras"
version = "0.1.0"
description = "Saras — E2E Agent Building Platform API"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",

    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",

    # Cache / real-time
    "redis[asyncio]>=5.2.0",

    # Analytics
    "duckdb>=1.1.0",

    # LLM
    "litellm>=1.55.0",
    "anthropic>=0.40.0",
    "openai>=1.58.0",

    # Data validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # YAML
    "pyyaml>=6.0.2",
    "jsonschema>=4.23.0",

    # Utilities
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "ulid-py>=1.1.0",

    # Eval — deterministic metrics
    "rouge-score>=0.1.2",
    "nltk>=3.8.0",
    "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.28.0",
    "anyio[trio]>=4.7.0",
    "testcontainers[postgres,redis]>=4.8.0",
    "factory-boy>=3.3.0",
    "faker>=33.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["saras"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
addopts = "--cov=saras --cov-report=term-missing --cov-fail-under=70"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
